fix(cellular): 修复单测编译错误, 恢复 build-asan 门禁与 MSVC 兼容
- tests: assert(org.is_compiled) -> is_compiled_ (成员名拼写错误, Release 下
  被 NDEBUG 掩盖, Debug/ASAN 下 error C2039 编译失败, 即 CI build-asan
  Build 阶段红掉的根因)
- cellular_genome: std::uniform_int_distribution<uint8_t> -> int + cast
  (标准 [rand.req.genl] 禁止 uint8_t, MSVC static_assert 直接报错;
  libstdc++ 宽容模板实参检查掩盖了该移植性问题)
- test 4 补强: forward() 为同步一拍延迟语义 (先汇聚上一拍输出再按拓扑序
  激发), 首拍动作细胞电位恒 0, 原断言只查 target_price, 买/卖/免疫锁路径
  从未被覆盖; 现在首拍断言 HOLD, 次拍真正断言 SELL_OPEN 决策与非空解释

本地 MSVC /O2 (assert 生效) 实测 7/7 通过, 前向传导 37.9 ns/次

+++ kun_quant/include/kun/cellular/cellular_genome.hpp
@@ -548,9 +548,9 @@ public:
         }
 
         std::uniform_real_distribution<double> dist_weight(-2.0, 2.0);
-        std::uniform_int_distribution<uint8_t> dist_port(0, 1);
+        std::uniform_int_distribution<int> dist_port(0, 1);
 
-        org.synapses.push_back({from_id, to_id, dist_port(rng_), dist_weight(rng_), true, 60.0f, -1.0f});
+        org.synapses.push_back({from_id, to_id, static_cast<uint8_t>(dist_port(rng_)), dist_weight(rng_), true, 60.0f, -1.0f});
         org.compile();
         return true;
     }
